requirements.txt
beautifulsoup4==4.12.3
blinker==1.8.2
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Flask==3.0.3
futures==3.0.5
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
logging==0.4.9.6
MarkupSafe==2.1.5
python-dotenv==1.0.1
requests==2.32.3
soupsieve==2.5
urllib3==2.2.1
Werkzeug==3.0.3
